+++ examples/irsa/outputs.tf
@@ -1,3 +1,4 @@
 output "aws_account_id" {
-  value = data.aws_caller_identity.current.account_id
+  description = "IAM AWS account id"
+  value       = data.aws_caller_identity.current.account_id
 }