+++ workers_launch_template.tf
@@ -2,11 +2,12 @@
 
 resource "aws_autoscaling_group" "workers_launch_template" {
   count = var.create_eks ? local.worker_group_launch_template_count : 0
+
   name_prefix = join(
     "-",
     compact(
       [
-        coalescelist(aws_eks_cluster.this[*].name, [""])[0],
+        local.cluster_name,
         lookup(var.worker_groups_launch_template[count.index], "name", count.index)
       ]
     )
@@ -219,7 +220,7 @@ resource "aws_autoscaling_group" "workers_launch_template" {
       [
         {
           "key" = "Name"
-          "value" = "${coalescelist(aws_eks_cluster.this[*].name, [""])[0]}-${lookup(
+          "value" = "${local.cluster_name}-${lookup(
             var.worker_groups_launch_template[count.index],
             "name",
             count.index,
@@ -227,7 +228,7 @@ resource "aws_autoscaling_group" "workers_launch_template" {
           "propagate_at_launch" = true
         },
         {
-          "key"                 = "kubernetes.io/cluster/${coalescelist(aws_eks_cluster.this[*].name, [""])[0]}"
+          "key"                 = "kubernetes.io/cluster/${local.cluster_name}"
           "value"               = "owned"
           "propagate_at_launch" = true
         },
@@ -289,7 +290,8 @@ resource "aws_autoscaling_group" "workers_launch_template" {
 
 resource "aws_launch_template" "workers_launch_template" {
   count = var.create_eks ? (local.worker_group_launch_template_count) : 0
-  name_prefix = "${coalescelist(aws_eks_cluster.this[*].name, [""])[0]}-${lookup(
+
+  name_prefix = "${local.cluster_name}-${lookup(
     var.worker_groups_launch_template[count.index],
     "name",
     count.index,
@@ -545,13 +547,18 @@ resource "aws_launch_template" "workers_launch_template" {
 
     tags = merge(
       {
-        "Name" = "${coalescelist(aws_eks_cluster.this[*].name, [""])[0]}-${lookup(
+        "Name" = "${local.cluster_name}-${lookup(
           var.worker_groups_launch_template[count.index],
           "name",
           count.index,
         )}-eks_asg"
       },
       var.tags,
+      {
+        for tag in lookup(var.worker_groups_launch_template[count.index], "tags", local.workers_group_defaults["tags"]) :
+        tag["key"] => tag["value"]
+        if tag["key"] != "Name" && tag["propagate_at_launch"]
+      }
     )
   }
 
@@ -560,7 +567,7 @@ resource "aws_launch_template" "workers_launch_template" {
 
     tags = merge(
       {
-        "Name" = "${coalescelist(aws_eks_cluster.this[*].name, [""])[0]}-${lookup(
+        "Name" = "${local.cluster_name}-${lookup(
           var.worker_groups_launch_template[count.index],
           "name",
           count.index,
@@ -569,6 +576,26 @@ resource "aws_launch_template" "workers_launch_template" {
       { for tag_key, tag_value in var.tags :
         tag_key => tag_value
         if tag_key != "Name" && !contains([for tag in lookup(var.worker_groups_launch_template[count.index], "tags", local.workers_group_defaults["tags"]) : tag["key"]], tag_key)
+      }
+    )
+  }
+
+  tag_specifications {
+    resource_type = "network-interface"
+
+    tags = merge(
+      {
+        "Name" = "${local.cluster_name}-${lookup(
+          var.worker_groups_launch_template[count.index],
+          "name",
+          count.index,
+        )}-eks_asg"
+      },
+      var.tags,
+      {
+        for tag in lookup(var.worker_groups_launch_template[count.index], "tags", local.workers_group_defaults["tags"]) :
+        tag["key"] => tag["value"]
+        if tag["key"] != "Name" && tag["propagate_at_launch"]
       }
     )
   }
@@ -597,14 +624,16 @@ resource "aws_launch_template" "workers_launch_template" {
 }
 
 resource "aws_iam_instance_profile" "workers_launch_template" {
-  count       = var.manage_worker_iam_resources && var.create_eks ? local.worker_group_launch_template_count : 0
-  name_prefix = coalescelist(aws_eks_cluster.this[*].name, [""])[0]
+  count = var.manage_worker_iam_resources && var.create_eks ? local.worker_group_launch_template_count : 0
+
+  name_prefix = local.cluster_name
   role = lookup(
     var.worker_groups_launch_template[count.index],
     "iam_role_id",
     local.default_iam_role_id,
   )
   path = var.iam_path
+
   tags = var.tags
 
   lifecycle {